add the Lockfile “decycler”
It breaks dependency cycles in the lockfiles. Took a while, but the few unit
tests I managed to implement seem to work. Would be a lot nicer with real
property tests. Especially the new Eq instance for MKMap, which I’m certainly
unsure about.

There is a pretty testing scheme using NamedFieldPuns, so that’s nice.