generate nix file with shortened package functions
We now have a pretty solid output that is 2/3 of the original `yarn.lock` for
our example.
Next step is to actually use the multi-keyed map. Hm, maybe that might even
bring the line-count *up* again, but should reduce file size (because package
definitions and thus hash sums will be shared.