relax `package.json` parsing by introducing warnings
Previously, every package that did not behave according to the spec of npm
`package.json` files was rejected by our library.
npm is so badly designed that it doesn’t even comply with its own
specifications, and has a complicated heuristics module dealing with missing or
wrongly typed fields and the like.

This patch tries to mirror a bit of that (uncalled for) leniency.