completely revamp the nix-side of yarn2nix & introduce nix-lib
The rather ad-hoc and muddled code from earlier is restructured, introducing a
sensible set of library functions (with nice signatures) to work with the
outputs generated by the `yarn2nix` utility.

Furthermore, it should now be possible to apply overrides to single packages and
their everything that depends on them (reverse dependencies). To achieve that,
the standard `nixpkgs` `fix/extend` mechanism is used.
This has the nice side-effect of getting rid of the ad-hoc manual fixpoints that
made the former code so very ad-hoc.