README/nix-lib: use local check out in nix-lib description
People trying out the library stumbled over the problem that `nix-lib` used the
`pkgs.haskellPackages.yarn2nix` by default, which is broken in current
unstable (because of the wrong megaparsec version).
Instead, the correct way to go for now is to also used the checked out
`yarn2nix` when calling `nix-lib`.

Changes the documentation to do just that and (temporarily) allows putting an
explicit `yarn2nix` package into the `nix-lib` which still falls back to the
`haskellPackages` version.