rewrite PackageKey parsing logic (bump 0.5.0)
Some “versions” were complete git+ssh links and contained `@` symbols.
Since we now parse scoped keys correctly (which assumes that the `@` can only
appear as very first caracter in a package name, we can safely split the version
on the second `@` (or first at for non-scoped) we see and the rest is version.