Add preliminary support for local packages
To support local packages (see commit for 0.6.0 in yarn-lock), we have
to add a new package function to the nix output.

For now it will only add the path directly, which has the drawback
that the nix expression is now required to be in the right
position (for relative paths). It should be changed to be user-defined
later (i.e. the user passes a function which takes the string of the
local path and returns a derivation to the package tarball).

We don’t use the tarball hash from the lockfile yet to verify the
integrity when importing into the nix store.