fix(nodejs/Package.hs): default to no (dev) dependencies if unparsable
In the node world you can't depend on anything, especially not that a
certain field of package.json is well-formed.

The dependencies field is sometimes malformed, probably in most cases it
is safe to default to {}. I have observed this in the wild with JSV 0.4.2
(https://github.com/garycourt/JSV) where dependencies is [] instead of
{} or being missing. I think packages with malformed dependency field
can't reasonably expect their dependencies to be installed correctly.

We only default to {} in cases where we can expect beyond a reasonable
doubt that there are no dependencies being expressed by the field:

* Empty array
* Scalars (Number, String, …)

We fail parsing if the field is malformed and seems to contain
something:

* Non-empty array
* Object that can't be parsed to Dependencies (i. e. is malformed)

We use this strategy for both dependencies and devDependencies.