feat(yarn2nix lib): lookup nixpkgs license sets by spdxId
This feature is in preparation of generating proper meta sets in
yarn2nix --template. We parse a licenses.json which is generated
from the nixpkgs lib/licenses.nix using writeText "licenses.json"
(builtins.toJSON lib.licenses) into a HashMap which uses the SPDX ID
as a key. Then we are able to lookup an SPDX ID in this HashMap
and return the nix expression representing the license referenced
by this SPDX ID in nixpkgs.

One limitation is that this approach is not dynamic, i. e. it
doesn't benefit from updates in nixpkgs, as the set rendered to
the template would be static instead of referencing the attribute
inside of lib.licenses. Thus all license sets generated would be as
up to date as the licenses.json used (probably as up to date as the
nixpkgs version yarn2nix was built with). However, this is actually
a feature: Any breaking changes to lib.licenses wouldn't break
existing yarn2nix templates' evaluation.