feat(yarn2nix): arg parsing using optparse-applicative
Unify runtime "configuration" mechanism: We now generate a RunConfig
from the command line and pass it to cli. This config is used to
determine which action to take (template / lock file) via RunMode
and contains additional information to influence runtime behavior
like runOffline.

The configuration can be passed to both actions if necessary, mandatory
action specific flags / variables could be hidden away in RunMode in the
future.