feat(yarn2nix): add proper license set to meta
If --license-data is given on the command line, lookup SPDX identifier
in that json file when generating a package template with --template.
If no license with said identifier can be found, the content of license
is added to meta.license as a string like before.

Possible improvement: Handle depreacated SPDX identifiers which seem
quite common (e. g. AGPL-3.0).