fix(Nodejs.License): workaround for missing shortName
Every member of lib.licenses should have a shortName attribute, this is
usually ensured a mapAttrs call in lib/licenses.nix. However currently
on nixpkgs master this is not the case due to a likely unintentional
change concerned with legacy license names.

Since it is almost trivial, we add a workaround for such cases which
exactly mirrors the mapAttrs call in lib/licenses.nix.

See also: https://github.com/NixOS/nixpkgs/pull/114416