chore(yarn2nix): Switch to aeson-2.0 (bump 0.10.0)
This was a minimally intrusive change to support the new aeson API.
Not much thought was spent on whether we want to use the
`Key`/`KeyMap` stuff internally or just convert to Text/Map directly.

The nixpkgs commit is from the GHC 9 migration PR at
https://github.com/NixOS/nixpkgs/pull/160733

We can remove the `easy-hls-nix` release and use the nixpkgs hls
instead, since it’s stable enough now.