chore(yarn2nix): move nix-tests to toplevel
Otherwise there was a conundrum where the tests would need a built
yarn2nix, but the source of yarn2nix would contain the nix test
sources, which means every time a nix-test changed the source of
yarn2nix would change and yarn2nix would have to be rebuilt.

This could certainly have been solved with fancy source filters, but
it was an boundary violation anyway, so moving them to toplevel solves
both problems.