feature(yarn2nix/Nodejs/License): rewrite license parsing
Instead of assuming we know exactly what fields the nixpkgs licenses
contain (which goes out of date every once in a while), just parse the
minimal necessary information we need (the `spdxID`) and then parse
just the type of the other fields. This way we get the same license
structs after generating the nix code again.

The only “bad apple” is our unfree license, since we assume to know
what the contents is, we will go out of sync with nixpkgs if any
fields change. That is probably not preventable.

Implementation-wise we switch to aeson-better-errors, since the parser
is so much nicer to use.